HW23 Extract pipelines from frameworks
